requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
filelock==3.18.0
fpdf==1.7.2
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
hf-xet==1.1.2
huggingface-hub==0.32.3
idna==3.10
importlib_metadata==7.2.1
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.41.0
networkx==3.4.2
numpy==1.26.4
packaging==23.2
pandas==2.2.3
pillow==10.4.0
protobuf==4.25.8
pyarrow==20.0.0
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.25.1
safetensors==0.5.3
six==1.17.0
smmap==5.0.2
streamlit==1.30.0
streamlit-drawable-canvas==0.9.3
sympy==1.14.0
tenacity==8.5.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
torchvision==0.22.0
tornado==6.5.1
tqdm==4.67.1
transformers==4.52.4
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
validators==0.35.0
zipp==3.22.0
